\name{checkerfun}
\alias{is.odd}
\alias{is.even}
\title{
Check for numbers for being either odd or even
}
\description{
Small helper functions to check number for either being odd or even.
}
\usage{
is.odd(k)
is.even(k)
}
\arguments{
  \item{k}{
An (integer) number
}
}
\value{
TRUE or FALSE, depending on whether k is odd or even (depending on the called function)
}
\author{
Karl Kugler <karl@eigenlab.net>
}
\examples{
is.odd(1)
is.even(1)
is.odd(2)
is.even(2)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ attribute }

