\name{topklists-package}
\alias{topklists-package}
\alias{TopKLists}
\alias{topklists}
\docType{package}
\title{
Inference, aggregation and visualization for top-k rank lists
}
\description{
More about what it does (maybe more than one line)
TODO: overall description
}
\details{
The package is split into three main moduls:
- TopKInference: TODO
- TopSpace: We provide functions to aggregate ranked lists based on three classes
of algorithms: Borda, Markov chain, and Cross-entropy Monte Carlo.
- TopKGraphics: Here we provide functions for visualizing the results from the other modules
}
\author{
Michael G. Schimek, Medical University of Graz, Austria (software
development and algorithms); Eva Budinska, Swiss Institute of Bioinformatics, Switzerland (implementation); Jie Ding, Harvard University, USA (implementation and algorithms); Peter Hall, The University of Melbourne, Australia (algorithms); Karl G. Kugler, UMIT, Austria (implementation and R-Forge management); Shili Lin, Ohio State University, USA (algorithms and implementation)
Maintainer: Michael G. Schimek <michael.schimek@medunigraz.at>
}
\seealso{
project homepage: \url{http://topklists.r-forge.r-project.org}
}
\references{
DeConde RP, Hawley S, Falcon S, Clegg N, Knudsen B, Etzioni R. Combining results of microarray experiments: a rank aggregation approach. Stat Appl Genet Mol Biol 2006, 5:15.

Lin, S., Ding, J. (2009) Integration of Ranked Lists via Cross Entropy Monte Carlo with Applications to mRNA and microRNA Studies. Biometrics, 65, 9-18. DOI: 10.1111/j.1541-0420.2008.01044.x. 

Lin, S. (2010) Space oriented rank-based data integration. Statistical Applications in Genetics and Molecular Biolog, 20, Article 20. DOI: 10.2202/1544-6115.1534. 

Lin, S. (2010) Rank aggregation methods. Wiley Interdisciplinary Reviews: Computational Statistics, 2, 555-570
}
\keyword{package}
