\name{TopKSpaceSampleInput}
\alias{TopKSpaceSampleInput}
\alias{input}
\alias{space}
\docType{data}
\title{
Sample input for TopKSpace functions
}
\description{
Sample input for \verb{TopKSpace} functions
}
\usage{data(TopKSpaceSampleInput)}
\format{
Two sets of lists for the demonstration of the usage of \verb{TopKSpace}-related functions
}
\examples{
data(TopKSpaceSampleInput)
str(input)
str(space)
}
\keyword{datasets}