\name{breast}
\docType{data}
\alias{breast}
\title{Sample data from breast cancer expression}
\description{
The example dataset comprises three lists of microarray results (differential gene
expression) from three breast cancer studies:
\enumerate{
 \item MicroArray Quality Control Phase II Project (2010) labeled MDACC
 \item Strong Time Dependence of the 76-Gene Prognostic Signature (2007) labeled TransBig
 \item A Clinically Relevant Gene Signature in Triple-Negative and Basal-Like Breast Cancer (2011) labeled Pusztai
}
Only genes (unique gene symbols) common to all studies are considered, therefore
each of the three lists has the length of N = 917.
}
\usage{breast}
\format{data.frame}
\source{
\enumerate{
 \item MDACC: \href{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20194}{GEO GSE20194}
 \item TransBig: \href{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=gse7390}{GEO GSE20194}
 \item Pusztai: \href{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE31519}{GEP GSE20194}
}
}

\keyword{datasets}

