\name{breast}
\docType{data}
\alias{breast}
\title{Sample data from breast cancer expression}
\description{
The example dataset comprises three lists of microarray results (differential gene
expression) from three breast cancer studies: 1) MicroArray Quality Control Phase II Project (2010) - data called MDACC, 2) Strong Time Dependence of the 76-Gene Prognostic Signature (2007) -
data called TransBig, 3) A Clinically Relevant Gene Signature in Triple-Negative and Basal-Like Breast Cancer (2011) - data called Pusztai
Only genes (unique gene symbols) common to all studies are considered, therefore
each of the three lists has the length of N = 917.
}
\usage{rivers}
\format{data.frame}
\source{MDACC: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE20194
TransBig: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=gse7390
Pusztai: http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE31519}

\keyword{datasets}

