\name{TopKGUISampleInput}
\alias{TopKGUISampleInput}
\alias{lists}
\docType{data}
\title{
Sample input for TopKGUI functions
}
\description{
Sample input for \verb{TopKGUI} functions
}
\usage{data(TopKGUISampleInput)}
\format{
Two sets of lists for demonstrating the usage of \verb{TopKGUI}-related functions
}
\examples{
data(TopKGUISampleInput)
str(lists)
}
