\name{TopKListsGUI}
\alias{TopKListsGUI}
\title{
Visualizations of TopKListsGUI Aggmap
}
\description{
Visualizations of TopKListsGUI Aggmap}
\usage{
TopKListsGUI(lists, autorange.delta = FALSE, override.errors = FALSE, aggmap.size = c(870, 440), aggmap.res = 100, venndiag.size = c(380, 420), venndiag.res = 70, gui.size = c(900, 810), directory = "TopKListsGUI_temp")
}
\arguments{
  \item{lists}{
}
  \item{autorange.delta}{
}
  \item{override.errors}{
}
  \item{aggmap.size}{
}
  \item{aggmap.res}{
}
  \item{venndiag.size}{
}
  \item{venndiag.res}{
}
  \item{gui.size}{
}
  \item{directory}{
}

}
\value{
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Eva Budinska
}
\note{
}
\seealso{
}
\examples{
##using a GUI
\dontrun{
data(TopKGUISampleInput)
TopKListsGUI(lists)
# Run for nu=10, delta from 2 to 10
}
}
\keyword{ graphics }
